convoluted local client-side http proxy
This solution works but it is hacky.  It allows us to manipulate the
Tunneled HTTP POST request by sending the TCP HTTP Tunnel to a HTTP
listener which then sends traffic to the downstream.

client -> Client Envoy TCP Proxy -> <HTTP tunnel> -> Client Envoy HTTP -> Server Envoy -> Server

This can work but it is a hack to say the least